Match UIPicture, UIListArrow and UISlider (#253)
* nearly match uipicture

* charikscale work

* charmeshhide work

* rndlight work

* uislider TU added

* uislider work

* more uislider work

* uislider nearly done

* nearly match uislider on msg

* match uislider onmsg

* nearly match uilistarrow

* uilistarrow draw work

* fix uilist ctor

* uilistarrow draw matched

* uilistarrow matched

* uislider matched

* match uipicture

* re-scope currently matched autogen weak funcs

---------

+++ src/system/utl/Loader.cpp
@@ -51,7 +51,7 @@ void LoadMgr::Init(){
     DataRegisterFunc("set_edit_mode", OnSetEditMode);
     DataRegisterFunc("set_loader_period", OnSetLoaderPeriod);
     DataRegisterFunc("sysplatform_sym", OnSysPlatformSym);
-    *DataVariable("sysplatform") = DataNode(mPlatform);
+    *DataVariable("sysplatform") = DataNode((int)mPlatform);
 }
 
 Loader* LoadMgr::ForceGetLoader(const FilePath& fp){
